use a config file to parse sprite information
newAnimationFromRow assumed that the each row of a sprite file was an
individual sprite, which is how I plan to format most of my work but not
ideal since shorter animations were left with blank frames.

newAnimationFromTable uses information from a config file to indicate
the start/end frames in the file, as well as eventually name each
sprite. unfortunately, for this method to work every frame in the file
has to have the same dimensions.

note: the HD sprite file is too big for Love to handle